Adding Test for MaxVertexDegree and AverageVertexDegree metrics
